refactor: move IsEmptyJSONObject to helpers/marshal.go
This function is a JSON utility, not a plan modifier. It belongs
alongside UnmarshalOptional in the helpers package.

Relates to #650
